Bgfx VK pm with photo attachment can not send
Reported on ccc-240. Previously API was changed but it was not
properly tested on real C3 wfsend-mails. Now this is fixed. The
reason it failed is because VK do not allow posting picture with
filename without extension. Cention mailattachments database
contain a localname field which do not has proper real file name
(without extension). This localname field was incorrectly used
previously to upload the picture. This commit will now use the
filename field of mailattachment database instead to make sure
it get pass VK API call.