Add poll attachment parsing and restructure geo
Wall post can include poll and it is considered as attachment.
This type of attachment was not parsed last time. This commit
enable the poll attachment parsing.

Geo struct is a new structure now which will be used in post
struct. Post struct now parse the Geo as pointer as not always
Geo will appear on Post. It only appear when user include map.
Using pointer can reduce the memory use during parsing Post
without map. With Geo struct, it is appended to Message struct
also as pm is allowed to send map info also.